% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{heath}
\alias{heath}
\title{Heathland vegetation data}
\format{
A tibble with 198 rows and 21 variables:
\describe{
  \item{site}{Site code}
  \item{year}{year of sampling}
  \item{species}{Species}
  \item{cover}{Percent cover}
  \item{freq1}{Presence/absence in subturf 1}
  ...
}
}
\source{
\url{https://www.uib.no/en/rg/EECRG/95156/landpress}
}
\description{
A dataset containing the cover and sub-turf presences of plants in two turfs on Norwegian coastal heathland recorded over four years.
}
