% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\alias{plot_subturf_grid}
\title{Make subturf grid}
\usage{
make_grid(grid, ncol, nrow = ncol)

plot_subturf_grid(grid_long)
}
\arguments{
\item{grid}{matrix; subturf ID of each subturf}

\item{ncol}{integer; number of columns}

\item{nrow}{integer; number of rows, defaults to `ncol` for a square grid}

\item{grid_long}{dataframe, output of `make_grid`}
}
\value{
`data.frame` with columns `.x`, `.y` and subturf

a ggplot object
}
\description{
Makes and optionally plots grid of sub-turfs.
}
\details{
use `ncol` and `nrow` if subturf IDs start at 1 in the 
top left and are read left-right. Use `grid` in other cases.
}
\examples{
#square
make_grid(ncol = 4)

#Rectangular
make_grid(ncol = 4, nrow = 5)

#snake
make_grid(grid = rbind(1:4, 8:5, 9:12, 16:13))

}
