% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_turf_plot.R
\name{make_turf_plot}
\alias{make_turf_plot}
\title{Subturf maps over time}
\usage{
make_turf_plot(
  data,
  grid_long,
  species,
  cover,
  year,
  subturf,
  title,
  site_id = "",
  turf_id = ""
)
}
\arguments{
\item{data}{Community data.frame in long form}

\item{grid_long}{grid from \code{make_grid}}

\item{species}{name of column with species data}

\item{cover}{name of column with cover data}

\item{year}{name of column with year data}

\item{subturf}{name of column with subturf IDs}

\item{title}{character; text for title if missing `site_id` and `turf_id` are used to make a title}

\item{site_id}{name of Site ID column for plot title}

\item{turf_id}{name of Turf ID column for plot title}
}
\value{
a ggplot object
}
\description{
Subturf maps over time
}
\details{
See vignette for examples
}
