% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_turf_plot_no_subturfs.R
\name{make_turf_plot_no_subturfs}
\alias{make_turf_plot_no_subturfs}
\title{Turf map over time}
\usage{
make_turf_plot_no_subturfs(data, year, species, cover, title)
}
\arguments{
\item{data}{Community data.frame in long form}

\item{year}{name of column with year data}

\item{species}{name of column with species data}

\item{cover}{name of column with cover data}

\item{title}{character; text for title}
}
\value{
a ggplot object
}
\description{
Plot species cover against time when there are no subturf information
}
\details{
This is just a wrapper for a simple ggplot, but made consistent with See vignette for examples
}
